# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Id": {
				"anyOf": [
					{
						"type": "number",
						"format": "double"
					},
					{
						"type": "string"
					}
				]
			},
...
			"Status": {
				"type": "string",
				"enum": [
					"ACTIVE",
					"INACTIVE",
					"DISABLED"
				]
			},
			"Role": {
				"type": "string",
				"enum": [
					"BASIC",
					"ADMIN"
				]
			},
			"User": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/Id"
					},
					"uuid": {
						"$ref": "#/components/schemas/UUID"
					},
					"name": {
						"type": "string"
					},
					"surname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"version": {
						"type": "number",
						"format": "double"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Company": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/Id"
					},
					"uuid": {
						"$ref": "#/components/schemas/UUID"
					},
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"province": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"fiscalCode": {
						"type": "string"
					},
					"vatCode": {
						"type": "string"
					},
					"registeredMunicipality": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"version": {
						"type": "number",
						"format": "double"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"LocalUnit": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/Id"
					},
					"uuid": {
						"$ref": "#/components/schemas/UUID"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"municipality": {
						"type": "string"
					},
					"province": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"itRea": {
						"type": "string"
					},
					"itCodiceAteco": {
						"type": "string"
					},
					"itAttivitaPrevalente": {
						"type": "string"
					},
					"itCciaa": {
						"type": "string"
					},
					"itIsArtigiana": {
						"type": "boolean"
					},
					"itIsAgricola": {
						"type": "boolean"
					},
					"itResponsabile": {
						"type": "string"
					},
					"itDatoreDiLavoro": {
						"type": "string"
					},
					"itDirigente": {
						"type": "string"
					},
					"itPreposto": {
						"type": "string"
					},
					"itRspp": {
						"type": "string"
					},
					"itAspp": {
						"type": "string"
					},
					"propertyId": {
						"$ref": "#/components/schemas/Id"
					},
					"companyId": {
						"$ref": "#/components/schemas/Id"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"version": {
						"type": "number",
						"format": "double"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Department": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/Id"
					},
					"uuid": {
						"$ref": "#/components/schemas/UUID"
					},
					"name": {
						"type": "string"
					},
					"localUnitId": {
						"$ref": "#/components/schemas/Id"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"version": {
						"type": "number",
						"format": "double"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Vehicle": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/Id"
					},
					"uuid": {
						"$ref": "#/components/schemas/UUID"
					},
					"name": {
						"type": "string"
					},
					"localUnitId": {
						"$ref": "#/components/schemas/Id"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"version": {
						"type": "number",
						"format": "double"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Equipment": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/Id"
					},
					"uuid": {
						"$ref": "#/components/schemas/UUID"
					},
					"name": {
						"type": "string"
					},
					"departmentId": {
						"$ref": "#/components/schemas/Id"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"version": {
						"type": "number",
						"format": "double"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"HR": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/Id"
					},
					"uuid": {
						"$ref": "#/components/schemas/UUID"
					},
					"name": {
						"type": "string"
					},
					"surname": {
						"type": "string"
					},
					"departmentId": {
						"$ref": "#/components/schemas/Id"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"version": {
						"type": "number",
						"format": "double"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Property": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/Id"
					},
					"uuid": {
						"$ref": "#/components/schemas/UUID"
					},
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"companyId": {
						"$ref": "#/components/schemas/Id"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"version": {
						"type": "number",
						"format": "double"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Document": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/Id"
					},
					"uuid": {
						"$ref": "#/components/schemas/UUID"
					},
					"name": {
						"type": "string"
					},
					"companyId": {
						"$ref": "#/components/schemas/Id"
					},
					"path": {
						"type": "string"
					},
					"refId": {
						"$ref": "#/components/schemas/Id"
					},
					"description": {
						"type": "string"
					},
					"moduleId": {
						"$ref": "#/components/schemas/Id"
					},
					"fileType": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"version": {
						"type": "number",
						"format": "double"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"name": "access_token"
			}
		}
	},
	"info": {
		"title": "Argos-IoT APIs",
		"version": "0.0.2",
		"description": "Restful APIs by Argos-IoT",
		"contact": {
			"name": "Marco Napoleone",
			"email": "napoleonemarco25@gmial.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/users/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "LogicDelete",
				"responses": {
					"204": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Logic delete a user",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
		"/companies": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Company"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all companies for a given user",
				"tags": [
					"Company"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Company"
								}
							}
						}
					}
				},
				"tags": [
					"Company"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Company"
							}
						}
					}
				}
			}
		},
		"/companies/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Company"
								}
							}
						}
					}
				},
				"tags": [
					"Company"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Company"
								}
							}
						}
					}
				},
				"tags": [
					"Company"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Company"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "LogicDelete",
				"responses": {
					"204": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Company"
								}
							}
						}
					}
				},
				"description": "Logic delete a company",
				"tags": [
					"Company"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
		"/companies/{id}/local-units": {
			"get": {
				"operationId": "GetAllLocalUnits",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LocalUnit"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Local unit",
					"Company"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
		"/companies/{id}/departments": {
			"get": {
				"operationId": "GetAllDepartments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Department"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Department",
					"Company"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
		"/companies/{id}/vehicles": {
			"get": {
				"operationId": "GetAllVehicles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Vehicle"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Vehicle",
					"Company"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
		"/companies/{id}/equipments": {
			"get": {
				"operationId": "GetAllEquipments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Equipment"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Equipment",
					"Company"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
		"/companies/{id}/hr": {
			"get": {
				"operationId": "GetAllHR",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/HR"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"HR",
					"Company"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
		"/companies/{id}/properties": {
			"get": {
				"operationId": "GetAllProperties",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Property"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Property",
					"Company"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
		"/companies/{id}/documents": {
			"get": {
				"operationId": "GetAllDocuments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Document"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Document",
					"Company"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
		"/companies/{id}/timetables": {
			"get": {
				"operationId": "GetAllTimetables",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Timetable"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Timetable",
					"Company"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
		"/local-units/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LocalUnit"
								}
							}
						}
					}
				},
				"tags": [
					"Local unit"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LocalUnit"
								}
							}
						}
					}
				},
				"tags": [
					"Local unit"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LocalUnit"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "LogicDelete",
				"responses": {
					"204": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LocalUnit"
								}
							}
						}
					}
				},
				"description": "Logic delete a localUnit",
				"tags": [
					"Local unit"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
...
		"/local-units/{id}/departments": {
			"get": {
				"operationId": "GetAllDepartments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Department"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Department",
					"Local unit"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
		"/local-units/{id}/vehicles": {
			"get": {
				"operationId": "GetAllVehicles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Vehicle"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Vehicle",
					"Local unit"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
...
		"/departments/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Department"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Department"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LocalUnit"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "LogicDelete",
				"responses": {
					"204": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Logic delete a department",
				"tags": [
					"Department"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
...
		"/departments/{id}/hr": {
			"get": {
				"operationId": "GetAllHR",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Department",
					"Department"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
		"/departments/{id}/equipment": {
			"get": {
				"operationId": "GetAllVehicles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Equipment",
					"Department"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
		"/hr/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"HR"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"HR"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LocalUnit"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "LogicDelete",
				"responses": {
					"204": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Logic delete a localUnit",
				"tags": [
					"HR"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
...
		"/vehicles/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Vehicle"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Vehicle"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LocalUnit"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "LogicDelete",
				"responses": {
					"204": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {},
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Logic delete a localUnit",
				"tags": [
					"Vehicle"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/Id"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}